+++ temba/flows/tasks.py
@@ -64,7 +64,7 @@ def check_flow_timeouts_task():
                 push_task(run.org_id, HANDLER_QUEUE, HANDLE_EVENT_TASK, task_payload)
 
                 queued_timeouts.set_queued([run])
-            except Exception:
+            except Exception:  # pragma: no cover
                 logger.error("Error queuing timeout task for run #%d" % run.id, exc_info=True)