+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+v3.0.139
+----------
+* Fix Macrokiosk JSON responses
+
+v3.0.138
+----------
+ * Migration to populate attachments field on old messages
+
+v3.0.137
+----------
+ * Don't assume event fires still exist in process_fire_events
+ * Add new Msg.attachments field to hold multiple attachments on an incoming message
+
+v3.0.136
+----------
+ * Fix scheduled broadcast text display
+
 v3.0.135
 ----------
  * Make 'only' keyword triggers ignore punctuation